requirements.txt
Django>=5.0
pymysql>=1.1.0
boto3>=1.40.0
django-storages[boto3]>=1.14.0
gunicorn>=21.2.0
python-decouple>=3.8
whitenoise>=6.0.0
cryptography>=40.0.0
pillow>=12.0.0